Made new function my_fun.
Best solutions can be founded in the phase of gradient descent. However, the values generated in gradient descent were not used for make new individuals. Therefore, even we found the best solution in gradient descent, it is never returned to coco correctly. To fix that, if we found the best solution in the phase of gradient descent, we would force to substitute it for self.all_best.
Also, improved some bad coding such as mutation mode was switched by boolean.